Updated PuTTY-CAC to 0.63 and Removed PKCS11 API
PuTTY-CAC is now updated to 0.63.  It fixes four security holes in 0.62
and before: vuln-modmul, vuln-signature-stringlen,
vuln-bignum-division-by-zero, private-key-not-wiped.  This version
includes the original interface for pageant with the added CAPI Certs
functionality.
The PKCS11 API has been removed from all applications in this PuTTY-CAC
Suite due to complications I was having with the code.  However, CAPI
support is fully functional which is the main premise behind PuTTY-CAC
anyways.  I have included compiled version of the PuTTY-CAC Suite that
can be found in the EXECUTABLES folder.  This includes all parts of the
PuTTY Suite except for PuTTYTEL.  If you choose to compile it yourself
you will need to use the MakeFile.vc as I did not update or support the
another MakeFile.* files.

The executables are compiled with VC 12 / VS 2013 Express.
These executables may not work on version of XP or Vista.  If this is
the case you will need to compile the code yourself using an older
version of VC / VS.